- made a storage class that wraps disk i/o for data files - recreated old behaviour in text_storage class - built binary_storage class: saves 60% of load time, adds 20% to file size.
next: gzip & others?

+++ src/common/gamecode/laws.c
@@ -2948,7 +2948,7 @@ age_building(building * b)
           effect.i = 100;
           /* the mage reactivates the circle */
           c = create_curse(mage, &rt->attribs, ct_astralblock,
-            sk, sk/2, effect, 0);
+            (float)sk, sk/2, effect, 0);
           ADDMSG(&r->msgs, msg_message("astralshield_activate", 
             "region unit", r, mage));
         }
